se extae la logica de locacion en un repositorio y se cambian nombres de paquetes en mayusculas
