+++ app/src/main/java/com/grupo7/moneychange/utils/PermissionChecker.kt
@@ -1,0 +1,28 @@
+package com.grupo7.moneychange.utils
+
+import android.app.Activity
+import com.karumi.dexter.Dexter
+import com.karumi.dexter.listener.PermissionDeniedResponse
+import com.karumi.dexter.listener.PermissionGrantedResponse
+import com.karumi.dexter.listener.single.BasePermissionListener
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.resume
+
+class PermissionChecker(private val activity: Activity, private val permission: String) {
+    suspend fun requestPermission(): Boolean =
+        suspendCancellableCoroutine { continuation ->
+            Dexter
+                .withActivity(activity)
+                .withPermission(permission)
+                .withListener(object : BasePermissionListener() {
+                    override fun onPermissionGranted(response: PermissionGrantedResponse?) {
+                        continuation.resume(true)
+                    }
+
+                    override fun onPermissionDenied(response: PermissionDeniedResponse?) {
+                        continuation.resume(false)
+                    }
+                }).check()
+        }
+
+}
