+++ app/src/main/java/com/grupo7/moneychange/viewModel/HistoryViewModel.kt
@@ -1,8 +1,8 @@
-package com.grupo7.moneychange.ViewModel
+package com.grupo7.moneychange.viewModel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
-import com.grupo7.moneychange.Repository.HistoryRepository
+import com.grupo7.moneychange.repository.HistoryRepository
 import com.grupo7.moneychange.data.entity.History
 
 /**
